+++ src/Services/DatabaseTools/AbstractDatabaseTool.php
@@ -11,14 +11,19 @@
 
 namespace Liip\TestFixturesBundle\Services\DatabaseTools;
 
+use BadMethodCallException;
 use Doctrine\Common\DataFixtures\Executor\AbstractExecutor;
 use Doctrine\Persistence\ManagerRegistry;
 use Doctrine\Persistence\ObjectManager;
 use Doctrine\DBAL\Connection;
+use InvalidArgumentException;
+use Liip\TestFixturesBundle\Event\FixtureEvent;
 use Liip\TestFixturesBundle\Services\DatabaseBackup\DatabaseBackupInterface;
 use Liip\TestFixturesBundle\Services\FixturesLoaderFactory;
+use ReflectionMethod;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
  * @author Aleksey Tupichenkov <alekseytupichenkov@gmail.com>
@@ -29,6 +34,9 @@ abstract class AbstractDatabaseTool
     const CACHE_METADATA_PARAMETER_NAME = 'liip_test_fixtures.cache_metadata';
 
     protected $container;
+
+    /** @var EventDispatcherInterface */
+    protected $eventDispatcher;
 
     protected $fixturesLoaderFactory;
 
@@ -77,6 +85,7 @@ abstract class AbstractDatabaseTool
     public function __construct(ContainerInterface $container, FixturesLoaderFactory $fixturesLoaderFactory)
     {
         $this->container = $container;
+        $this->eventDispatcher = $container->get('event_dispatcher');
         $this->fixturesLoaderFactory = $fixturesLoaderFactory;
     }
 
@@ -137,13 +146,13 @@ abstract class AbstractDatabaseTool
     abstract public function loadFixtures(array $classNames = [], bool $append = false): AbstractExecutor;
 
     /**
-     * @throws \BadMethodCallException
+     * @throws BadMethodCallException
      */
     public function loadAliceFixture(array $paths = [], bool $append = false): array
     {
         $persisterLoaderServiceName = 'fidry_alice_data_fixtures.loader.doctrine';
         if (!$this->container->has($persisterLoaderServiceName)) {
-            throw new \BadMethodCallException('theofidry/alice-data-fixtures must be installed to use this method.');
+            throw new BadMethodCallException('theofidry/alice-data-fixtures must be installed to use this method.');
         }
 
         if (false === $append) {
@@ -163,7 +172,7 @@ abstract class AbstractDatabaseTool
     /**
      * Locate fixture files.
      *
-     * @throws \InvalidArgumentException if a wrong path is given outside a bundle
+     * @throws InvalidArgumentException if a wrong path is given outside a bundle
      */
     protected function locateResources(array $paths): array
     {
@@ -174,7 +183,7 @@ abstract class AbstractDatabaseTool
         foreach ($paths as $path) {
             if ('@' !== $path[0]) {
                 if (!file_exists($path)) {
-                    throw new \InvalidArgumentException(sprintf('Unable to find file "%s".', $path));
+                    throw new InvalidArgumentException(sprintf('Unable to find file "%s".', $path));
                 }
                 $files[] = $path;