+++ src/Services/DatabaseTools/ORMDatabaseTool.php
@@ -19,6 +19,11 @@ use Doctrine\DBAL\DriverManager;
 use Doctrine\DBAL\Platforms\MySqlPlatform;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Tools\SchemaTool;
+use Liip\TestFixturesBundle\Event\FixtureEvent;
+use Liip\TestFixturesBundle\Event\PostFixtureBackupRestoreEvent;
+use Liip\TestFixturesBundle\Event\PreFixtureBackupRestoreEvent;
+use Liip\TestFixturesBundle\Event\ReferenceSaveEvent;
+use Liip\TestFixturesBundle\LiipTestFixturesEvents;
 
 /**
  * @author Aleksey Tupichenkov <alekseytupichenkov@gmail.com>
@@ -115,11 +120,15 @@ class ORMDatabaseTool extends AbstractDatabaseTool
                 $this->om->flush();
                 $this->om->clear();
 
-                $this->testCase->preFixtureBackupRestore($this->om, $referenceRepository, $backupService->getBackupFilePath());
+                $event = new PreFixtureBackupRestoreEvent($this->om, $referenceRepository, $backupService->getBackupFilePath());
+                $this->eventDispatcher->dispatch($event, LiipTestFixturesEvents::PRE_FIXTURE_BACKUP_RESTORE);
+
                 $executor = $this->getExecutor($this->getPurger());
                 $executor->setReferenceRepository($referenceRepository);
                 $backupService->restore($executor, $this->excludedDoctrineTables);
-                $this->testCase->postFixtureBackupRestore($backupService->getBackupFilePath());
+
+                $event = new PostFixtureBackupRestoreEvent($backupService->getBackupFilePath());
+                $this->eventDispatcher->dispatch($event, LiipTestFixturesEvents::POST_FIXTURE_BACKUP_RESTORE);
 
                 return $executor;
             }
@@ -141,7 +150,8 @@ class ORMDatabaseTool extends AbstractDatabaseTool
             }
         }
 
-        $this->testCase->postFixtureSetup();
+        $event = new FixtureEvent();
+        $this->eventDispatcher->dispatch($event, LiipTestFixturesEvents::POST_FIXTURE_SETUP);
 
         $executor = $this->getExecutor($this->getPurger());
         $executor->setReferenceRepository($referenceRepository);
@@ -155,9 +165,12 @@ class ORMDatabaseTool extends AbstractDatabaseTool
         $executor->execute($loader->getFixtures(), true);
 
         if ($backupService) {
-            $this->testCase->preReferenceSave($this->om, $executor, $backupService->getBackupFilePath());
+            $event = new ReferenceSaveEvent($this->om, $executor, $backupService->getBackupFilePath());
+            $this->eventDispatcher->dispatch($event, LiipTestFixturesEvents::PRE_REFERENCE_SAVE);
+
             $backupService->backup($executor);
-            $this->testCase->postReferenceSave($this->om, $executor, $backupService->getBackupFilePath());
+
+            $this->eventDispatcher->dispatch($event, LiipTestFixturesEvents::POST_REFERENCE_SAVE);
         }
 
         return $executor;